feat: Implementeing Evnetual Consistency via an Eventual Consistency Middleware

+++ GymManagement.Application/Subscriptions/Events/SubscriptionDeletedEventhandler.cs
@@ -13,7 +13,12 @@ public class SubscriptionDeletedEventHandler(ISubscriptionsRepository subscripti
 
     public async Task Handle(SubscriptionDeletedEvent notification, CancellationToken cancellationToken)
     {
-        var subscription = await _subscriptionsRepository.GetByIdAsync(notification.SubscriptionId) ?? throw new InvalidOperationException("Subscription not found");
+        var subscription = await _subscriptionsRepository.GetByIdAsync(notification.SubscriptionId);
+
+        if (subscription is null)
+        {
+            throw new InvalidOperationException("Subscription is not found");
+        }
 
         await _subscriptionsRepository.DeleteSubscriptionAsync(subscription);
         await _unitOfWork.CommitChangesAsync();
